Move sampling functions to renderer/utility
