+++ src/appleseed/renderer/modeling/postprocessingstage/bloompostprocessingstage.cpp
@@ -33,6 +33,7 @@
 #include "renderer/global/globallogger.h"
 #include "renderer/modeling/frame/frame.h"
 #include "renderer/modeling/postprocessingstage/postprocessingstage.h"
+#include "renderer/utility/rgbcolorsampling.h"
 
 // appleseed.foundation headers.
 #include "foundation/containers/dictionary.h"
@@ -105,194 +106,6 @@ namespace
             m_debug_threshold = m_params.get_optional("debug_threshold", false, context);
 
             return true;
-        }
-
-        //
-        // Simple image color sampling.
-        //
-
-        // Bilinear interpolation of colors.
-        static Color3f blerp(
-            const Image&        image,
-            const std::size_t   x0,
-            const std::size_t   y0,
-            const std::size_t   x1,
-            const std::size_t   y1,
-            const float         fx,
-            const float         fy)
-        {
-            // Retrieve the four surrounding pixels.
-            Color3f c00, c10, c01, c11;
-            image.get_pixel(x0, y0, c00);
-            image.get_pixel(x1, y0, c10);
-            image.get_pixel(x0, y1, c01);
-            image.get_pixel(x1, y1, c11);
-
-            // Compute weights.
-            const float wx1 = fx - x0;
-            const float wy1 = fy - y0;
-            const float wx0 = 1.0f - wx1;
-            const float wy0 = 1.0f - wy1;
-
-            // Return the weighted sum.
-            const Color3f result =
-                c00 * wx0 * wy0 +
-                c10 * wx1 * wy0 +
-                c01 * wx0 * wy1 +
-                c11 * wx1 * wy1;
-
-            return result;
-        }
-
-        // Returns the weighted average of the four pixels closest to the image coordinate (fx, fy).
-        static Color3f box_sample(const Image& image, const float fx, const float fy)
-        {
-            const std::size_t x0 = truncate<std::size_t>(fx);
-            const std::size_t y0 = truncate<std::size_t>(fy);
-            const std::size_t x1 = std::min<std::size_t>(x0 + 1, image.properties().m_canvas_width - 1);
-            const std::size_t y1 = std::min<std::size_t>(y0 + 1, image.properties().m_canvas_height - 1);
-
-            return blerp(image, x0, y0, x1, y1, fx, fy);
-        }
-
-        // Returns the weighted average of the four pixels closest to the image coordinate (fx, fy).
-        // Samples the edge-most pixels when coordinates are outside the image (i.e. texture clamping).
-        static Color3f clamped_box_sample(const Image& image, const float fx, const float fy)
-        {
-            const std::size_t width = image.properties().m_canvas_width;
-            const std::size_t height = image.properties().m_canvas_height;
-
-            // Compute to which region the coordinate (fx, fy) belongs, similar to Cohen–Sutherland's algorithm.
-            const int x_region = (fx < 0.0f) ? -1 : (fx > width - 1.0f) ? 1 : 0;
-            const int y_region = (fy < 0.0f) ? -1 : (fy > height - 1.0f) ? 1 : 0;
-
-            // There are 9 regions relative to the image, with 3 base cases:
-            //   -+ 0+ ++    inside: 00
-            //   -0 00 +0    outside lateral: 0+, -0, +0, 0-
-            //   -- 0- +-    outside diagonal: -+, ++, --, +-
-
-            // Thus, we can easily check the region of (fx, fy) and clamp it to be inside.
-            const std::size_t x = static_cast<std::size_t>((x_region == -1) ? 0 : width - 1);
-            const std::size_t y = static_cast<std::size_t>((y_region == -1) ? 0 : height - 1);
-
-            if (x_region * y_region != 0)
-            {
-                Color3f corner_pixel;
-                image.get_pixel(x, y, corner_pixel);
-
-                return corner_pixel;
-            }
-
-            const std::size_t x0 = truncate<std::size_t>(fx);
-            const std::size_t y0 = truncate<std::size_t>(fy);
-            const std::size_t x1 = std::min<std::size_t>(x0 + 1, width - 1);
-            const std::size_t y1 = std::min<std::size_t>(y0 + 1, height - 1);
-
-            if (x_region != 0)
-            {
-                Color3f top_pixel, bottom_pixel;
-                image.get_pixel(x, y0, bottom_pixel);
-                image.get_pixel(x, y1, top_pixel);
-
-                return lerp(bottom_pixel, top_pixel, fy - y0);
-            }
-            else if (y_region != 0)
-            {
-                Color3f left_pixel, right_pixel;
-                image.get_pixel(x0, y, left_pixel);
-                image.get_pixel(x1, y, right_pixel);
-
-                return lerp(left_pixel, right_pixel, fx - x0);
-            }
-            else // inside the image
-            {
-                return blerp(image, x0, y0, x1, y1, fx, fy);
-            }
-        }
-
-        //
-        // Sampling filter from Masaki Kawase's GDC2003 Presentation: "Frame Buffer Postprocessing Effects in DOUBLE-S.T.E.A.L (Wreckless)".
-        //
-
-        static Color3f kawase_sample(
-            const Image& image,
-            const std::size_t x,
-            const std::size_t y,
-            const std::size_t offset)
-        {
-            const float fx = static_cast<float>(x);
-            const float fy = static_cast<float>(y);
-            const float off = static_cast<float>(offset) + 0.5f;
-
-            // Since each corner sample is an average of 4 values, 16 pixels are used in total.
-            Color3f top_left = clamped_box_sample(image, fx - off, fy + off);
-            Color3f top_right = clamped_box_sample(image, fx + off, fy + off);
-            Color3f bottom_left = clamped_box_sample(image, fx - off, fy - off);
-            Color3f bottom_right = clamped_box_sample(image, fx + off, fy - off);
-
-            // Average sample colors.
-            return 0.25f * (top_left + top_right + bottom_left + bottom_right);
-        }
-
-        //
-        // Sampling filters from Marius Bjørge's SIGGRAPH2015 Presentation: "Bandwidth-Efficient Rendering".
-        //
-
-        // Note: corner_x and corner_y should be half-integer values.
-        static Color3f dual_filter_downsample(
-            const Image& image,
-            const float corner_x,
-            const float corner_y,
-            const std::size_t offset = 1)
-        {
-            const float fx = corner_x;
-            const float fy = corner_y;
-            const float half_off = 0.5f * static_cast<float>(offset);
-
-            Color3f center = clamped_box_sample(image, fx, fy);
-            Color3f top_left = clamped_box_sample(image, fx - half_off, fy + half_off);
-            Color3f top_right = clamped_box_sample(image, fx + half_off, fy + half_off);
-            Color3f bottom_left = clamped_box_sample(image, fx - half_off, fy - half_off);
-            Color3f bottom_right = clamped_box_sample(image, fx + half_off, fy - half_off);
-
-            return (
-                4.0f * center
-                + top_left + top_right + bottom_left + bottom_right)
-                / 8.0f;
-        }
-
-        static Color3f dual_filter_upsample(
-            const Image& image,
-            const std::size_t center_x,
-            const std::size_t center_y,
-            const std::size_t offset = 1)
-        {
-            const float fx = static_cast<float>(center_x);
-            const float fy = static_cast<float>(center_y);
-            const float off = static_cast<float>(offset);
-            const float half_off = 0.5f * off;
-
-            Color3f top_left = clamped_box_sample(image, fx - half_off, fy + half_off);
-            Color3f top_right = clamped_box_sample(image, fx + half_off, fy + half_off);
-            Color3f bottom_left = clamped_box_sample(image, fx - half_off, fy - half_off);
-            Color3f bottom_right = clamped_box_sample(image, fx + half_off, fy - half_off);
-
-            // Sample the edge-most pixel when the coordinate is outside the image (i.e. texture clamping).
-            const std::size_t top_y = static_cast<std::size_t>(std::min(fy + off, image.properties().m_canvas_height - 1.0f));
-            const std::size_t left_x = static_cast<std::size_t>(std::max(fx - off, 0.0f));
-            const std::size_t right_x = static_cast<std::size_t>(std::min(fx + off, image.properties().m_canvas_width - 1.0f));
-            const std::size_t bottom_y = static_cast<std::size_t>(std::max(fy - off, 0.0f));
-
-            Color3f top, left, right, bottom;
-            image.get_pixel(center_x, top_y, top);
-            image.get_pixel(left_x, center_y, left);
-            image.get_pixel(right_x, center_y, right);
-            image.get_pixel(center_x, bottom_y, bottom);
-
-            return (
-                top + left + right + bottom
-                + 2.0f * (top_left + top_right + bottom_left + bottom_right))
-                / 12.0f;
         }
 
         //
